scripts/requirements.txt
epitran
gtts
pyyaml
requests
